# Conflicts:
#	Docker/centralserver/Dockerfile
#	Docker/centralserver/files/cs-xroad.conf
#	Docker/securityserver/Dockerfile
#	Docker/securityserver/init_context.sh
#	Docker/xrd-dev-stack/.env
#	Docker/xrd-dev-stack/.env.local
#	Docker/xrd-dev-stack/hurl-src/vars.env
#	src/addons/messagelog/messagelog-addon/core/src/main/java/ee/ria/xroad/proxy/messagelog/LogManager.java
#	src/central-server/admin-service/globalconf-generator/src/main/java/org/niis/xroad/cs/admin/globalconf/generator/SharedParameters.java
#	src/central-server/admin-service/infra-api-rest/src/test/java/org/niis/xroad/cs/admin/rest/api/converter/AbstractDtoConverterTest.java
#	src/central-server/admin-service/infra-jpa/src/main/resources/liquibase/centerui-changelog.xml
#	src/common/common-core/src/main/java/ee/ria/xroad/common/SystemProperties.java
#	src/common/common-globalconf/src/main/java/ee/ria/xroad/common/conf/globalconf/SharedParameters.java
#	src/common/common-globalconf/src/main/java/ee/ria/xroad/common/conf/globalconf/SharedParametersV2Converter.java
#	src/common/common-globalconf/src/main/java/ee/ria/xroad/common/conf/globalconf/SharedParametersV3Converter.java
#	src/common/common-globalconf/src/test/java/ee/ria/xroad/common/conf/globalconf/SharedParametersV3ToXmlConverterTest.java
#	src/common/common-int-test/src/main/resources/signer-container-files/Dockerfile
#	src/common/common-verifier/src/main/java/ee/ria/xroad/common/conf/globalconf/GlobalConfImpl.java
#	src/gradle/libs.versions.toml
#	src/proxy/src/main/java/ee/ria/xroad/proxy/clientproxy/AbstractClientProxyHandler.java
#	src/proxy/src/main/java/ee/ria/xroad/proxy/clientproxy/ClientSoapMessageProcessor.java
#	src/security-server/e2e-test/src/intTest/resources/behavior/0100-ss-proxy.feature

+++ src/common/common-globalconf/src/main/java/ee/ria/xroad/common/conf/globalconf/PrivateParametersV2Converter.java
@@ -36,7 +36,7 @@ public class PrivateParametersV2Converter {
     PrivateParameters convert(PrivateParametersType source) {
         var target = new PrivateParameters();
         target.setInstanceIdentifier(source.getInstanceIdentifier());
-        target.setTimeStampingIntervalSeconds(source.getTimeStampingIntervalSeconds());
+        target.setTimeStampingIntervalSeconds(source.getTimeStampingIntervalSeconds().intValue());
 
         if (source.getManagementService() != null) {
             var managementService = new PrivateParameters.ManagementService();
